requirements.txt
opencv-python
streamlit
ultralytics
numpy
pillow